botanicalbuddy/requirements.txt
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
django-filter==24.3
setuptools==75.6.0
pydantic-ai-slim[openai]==0.0.11
pydantic==2.10
django==5.1.4
python-dotenv==1.0.1
django-filter==24.3
psycopg[BINARY]==3.2.3
requests==2.32.3
numpy==2.2.0
scikit-learn==1.5.2
openai==1.57.3
pgvector==0.3.6
spacy==3.8.3
pillow==11.0.0